backend/requirements.txt
# ============================================
# AI Event Agent — Backend Dependencies
# ============================================

# Agent Framework
crewai>=0.80.0
crewai-tools>=0.12.0

# LLM Integration
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# Web Scraping
crawl4ai>=0.4.0

# Search
duckduckgo-search>=6.0.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0

# Deduplication
rapidfuzz>=3.0.0

# Config
python-dotenv>=1.0.0

# Auth + Scheduler + PDF
PyJWT>=2.8.0
apscheduler>=3.10.4
fpdf2>=2.7.9

# Utilities
httpx>=0.27.0
